Fix test
This was actually a poor test, functionality is working as expected.
We used to be able to cheat and get the package info as a public user,
but now we can't. This gets that info logged in, then hits the
resource route as an anonymous user and validates the download.