TEMP: disable auto-deploy to production
After manual deployment to production, we will use CI to deploy to staging to make sure there are no discrepencies between our manual and auto deployments.  If it all goes well, then we'll make a new PR to auto deploy to production.  The motivation for this is that sometimes CI has issues communicating with a manually pushed app or vice-versa... there will be no code changes, but things might change outside of code.